Refactor: extract shared UI components
